tests/wpt/tests/shared-storage/run-url-selection-operation-without-add-module.tentative.https.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
'use strict';

for (const resolve_to_config of [true, false]) {
  promise_test(async t => {
    return promise_rejects_dom(t, "OperationError",
      sharedStorage.selectURL(
          "test-url-selection-operation",
          [{url: "/shared-storage/resources/frame0.html"}],
          {resolveToConfig: resolve_to_config, keepAlive: true}));
  }, 'selectURL() without addModule, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
      try {
        await sharedStorage.selectURL(
            "test-url-selection-operation",
            [{url: "1"}, {url: "2"}, {url: "3"}, {url: "4"}, {url: "5"},
             {url: "6"}, {url: "7"}, {url: "8"}, {url: "9"}],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
      assert_equals(e.message, 'Length of the \"urls\" parameter is not valid.');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with urls array length too big, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
        await sharedStorage.selectURL(
          "test-url-selection-operation",
          [],
          {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
      assert_equals(e.message, 'Length of the \"urls\" parameter is not valid.');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with empty urls array, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
        await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                reportingMetadata: {
                    'click': "https://google.com"
                }
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'TypeError');
        assert_equals(true, e.message.startsWith(
            'Failed to execute \'selectURL\' on \'SharedStorage\': ' +
            'Failed to read the \'url\' property from ' +
            '\'SharedStorageUrlWithMetadata\':') &&
            e.message.endsWith('Failed to read the \'url\' property from ' +
            '\'SharedStorageUrlWithMetadata\': Required member is undefined.'));
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with missing url, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
        await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "https://#"
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message,
                      'The url \"https://#\" is invalid.');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with invalid url, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
        await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "/shared-storage/resources/frame0.html",
                reportingMetadata: {
                    'click': "https://#"
                }
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message,
                      'The metadata for the url at index 0 has an invalid ' +
                      'or non-HTTPS report_url parameter \"https://#\".');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with invalid report url ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
        await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "/shared-storage/resources/frame0.html",
                reportingMetadata: {
                    'click': "http://google.com"
                }
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message, 'The metadata for the url at index 0 has an ' +
            'invalid or non-HTTPS report_url parameter \"http://google.com\".');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with http report url, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
        await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "/shared-storage/resources/frame0.html",
                reportingMetadata: {}
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message, 'selectURL could not get reportingMetadata ' +
                                 'object attributes');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with invalid reportingMetadata ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));
}

</script>
</body>
